requirements.txt
fastapi
uvicorn[standard]
pydantic[email]
mysql-connector-python
passlib[bcrypt]
python-dotenv
python-jose[cryptography]
python-multipart